generate menus server-side
commit from jmaessen:

Generate menus on the server side using mobilize_menu_render_filter doing a
second fetch of the HTML in the background and running mobilize_label_filter and
mobilize_menu_filter on the results.  Inject the navigation bar on the server
side and adapt the mobilize_nav.js to properly handle the prsence of these
elements on the rewritten page.

+++ pagespeed/kernel/base/null_statistics.h
@@ -36,7 +36,7 @@ class NullStatisticsVariable {
   ~NullStatisticsVariable() {}
   void Set(int64 value) { }
   int64 Get() const { return 0; }
-  int64 AddHelper(int delta) const { return 0; }
+  int64 AddHelper(int64 delta) const { return 0; }
   StringPiece GetName() const { return StringPiece(NULL); }
 
  private:
